+++ tool/templates/riverpod_mutation.dart
@@ -19,16 +19,16 @@ abstract class MutationRef {
 
   void invalidate(ProviderOrFamily provider) => _container.invalidate(provider);
 
-  T read<T>(ProviderListenable<T> provider) => _container.read(provider);
+  T read<T>(ProviderListenable<T> provider);
 
-  T refresh<T>(Refreshable<T> provider) => _container.refresh(provider);
+  T refresh<T>(Refreshable<T> provider);
 
   void updateProgress(double value);
 }
 
-extension MutationNotififierExtension on ConsumerScope {
+extension MutationNotififierExtension on SourceWidgetRef {
   MutationNotifier<A, R> mutation<A, R>(Future<R> Function(MutationRef ref, A arg) mutator) {
-    final mutation = MutationNotifier<A, R>(() => ref.container, mutator);
+    final mutation = MutationNotifier<A, R>(this, mutator);
     onDispose(mutation.dispose);
     return mutation;
   }
@@ -39,10 +39,10 @@ typedef DataMutationListener<Result> = FutureOr<void> Function(Result result);
 typedef ResultMutationListener<Result> = FutureOr<void> Function(Object? error, Result? result);
 
 class MutationNotifier<TArg, TResult> extends SourceNotifier<MutationState<TResult>> {
-  final ProviderContainer Function() _readContainer;
+  final SourceWidgetRef _ref;
   final Future<TResult> Function(MutationRef ref, TArg arg) _mutator;
 
-  MutationNotifier(this._readContainer, this._mutator) : super(IdleMutation<TResult>());
+  MutationNotifier(this._ref, this._mutator) : super(IdleMutation<TResult>());
 
   void call(
     TArg arg, {
@@ -61,7 +61,7 @@ class MutationNotifier<TArg, TResult> extends SourceNotifier<MutationState<TResu
 
     state = state.toLoading(arg: arg);
 
-    final ref = _MutationRef(_readContainer(), this, arg);
+    final ref = _MutationRef(_ref.container, this, arg);
     try {
       final result = await _mutator(ref, arg);
       ref.dispose();
@@ -328,12 +328,29 @@ class _MutationRef<TArg> extends MutationRef {
   MutationNotifier? _notifier;
   final TArg _arg;
 
+  final _subscriptions = <ProviderListenable, ProviderSubscription>{};
+
   _MutationRef(super._container, this._notifier, this._arg);
+
+  @override
+  T read<T>(ProviderListenable<T> provider) {
+    _subscriptions.putIfAbsent(provider, () => _container.listen(provider, (_, _) {}));
+    return _container.read(provider);
+  }
+
+  @override
+  T refresh<T>(Refreshable<T> provider) {
+    _subscriptions.putIfAbsent(provider, () => _container.listen(provider, (_, _) {}));
+    return _container.refresh(provider);
+  }
 
   @override
   void updateProgress(double value) => _notifier?._updateProgress(_arg, value);
 
   void dispose() {
+    for (final subscription in _subscriptions.values) {
+      subscription.close();
+    }
     _notifier = null;
   }
 }